requirements.txt
absl-py==1.0.0
aioredis==1.3.1
APScheduler==3.9.1
argon2-cffi==20.1.0
asgiref==3.2.10
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
blinker==1.4
boto3==1.17.44
botocore==1.20.112
bs4==0.0.1
cachetools==5.0.0
certifi==2020.6.20
cffi==1.14.0
channels==3.0.5
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
config==0.5.1
constantly==15.1.0
construct==2.5.3
cryptography==38.0.1
cycler==0.10.0
daphne==3.0.2
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.6.0
Deprecated==1.2.13
Django==3.0.8
django-allauth==0.42.0
django-background-tasks==1.2.5
django-bootstrap-v5==1.0.0
django-compat==1.0.15
django-cors-headers==3.13.0
django-crispy-forms==1.9.1
django-db-file-storage==0.5.5
django-debug-toolbar==2.2
django-environ==0.4.5
django-markdownify==0.9.0
django-model-utils==4.0.0
django-nested-admin==3.4.0
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.0
entrypoints==0.3
Flask==2.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
haystack==0.42
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.11.3
imutils==0.5.4
incremental==21.3.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mlxtend==0.19.0
msgpack==1.0.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.6
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.0
olympia==0.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandas-datareader==0.10.0
pandocfilters==1.4.3
parso==0.8.2
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.3
python-monkey-business==1.0.0
python-ptrace==0.9.8
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
redis==4.3.4
regex==2021.11.10
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.3.7
scikit-learn==0.24.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==21.1.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.36
sqlparse==0.3.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.2.0
timeago==1.0.15
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.25.9
utils==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.0
whitenoise==6.2.0
widgetsnbextension==3.5.1
wrapt==1.14.0
WTForms==3.0.1
xgboost==1.5.2
zipp==3.7.0
zope.interface==5.4.0
